requirements.txt
functions-framework==3.*
 
annotated-types==0.7.0
 
anyio==4.9.0
 
attrs==25.3.0
 
Authlib==1.6.1
 
blinker==1.9.0
 
cachetools==5.5.2
 
mem0ai
 
certifi==2025.7.14
 
cffi==1.17.1
 
charset-normalizer==3.4.2
 
click==8.2.1
 
cloudpickle==3.1.1
 
colorama==0.4.6
 
cryptography==45.0.5
 
docker==7.1.0
 
docstring_parser==0.17.0
 
fastapi==0.116.1
 
Flask==3.1.1

google-cloud-pubsub

google-cloud-tasks>=2.16.0
 
fpdf2==2.7.9
 
google-adk
 
google-api-core==2.25.1
 
google-api-python-client==2.177.0
 
google-auth==2.40.3
 
google-auth-httplib2==0.2.0
 
google-cloud-aiplatform==1.105.0
 
google-cloud-appengine-logging==1.6.2
 
google-cloud-audit-log==0.3.2
 
google-cloud-bigquery==3.35.0
 
google-cloud-core==2.4.3
 
google-cloud-logging==3.12.1
 
google-cloud-resource-manager==1.14.2
 
google-cloud-secret-manager==2.24.0
 
google-cloud-speech==2.33.0
 
google-cloud-storage==2.19.0
 
google-cloud-trace==1.16.2
 
google-crc32c==1.7.1
 
google-genai==1.27.0

google-generativeai

google-resumable-media==2.7.2
 
googleapis-common-protos==1.70.0
 
graphviz==0.21
 
greenlet==3.2.3
 
grpc-google-iam-v1==0.14.2
 
grpcio==1.73.1
 
grpcio-status
 
h11==0.16.0
 
httpcore==1.0.9
 
httplib2==0.22.0
 
httpx==0.28.1
 
httpx-sse==0.4.1
 
idna==3.10
 
importlib_metadata==8.7.0
 
itsdangerous==2.2.0
 
Jinja2==3.1.6
 
jsonschema==4.25.0
 
jsonschema-specifications==2025.4.1
 
MarkupSafe==3.0.2
 
mcp==1.12.1
 
nest-asyncio==1.6.0
 
numpy==2.3.1
 
opentelemetry-api==1.35.0
 
opentelemetry-exporter-gcp-trace==1.9.0
 
opentelemetry-resourcedetector-gcp==1.9.0a0
 
opentelemetry-sdk==1.35.0
 
opentelemetry-semantic-conventions==0.56b0
 
packaging
 
proto-plus==1.26.1
 
protobuf
 
pyasn1==0.6.1
 
pyasn1_modules==0.4.2
 
pycparser==2.22
 
pydantic==2.11.7
 
pydantic-settings==2.10.1
 
pydantic_core==2.33.2
 
pyparsing==3.2.3
 
python-dateutil==2.9.0.post0
 
python-dotenv==1.1.1
 
python-multipart==0.0.20
 
PyYAML==6.0.2
 
referencing==0.36.2
 
requests==2.32.4
 
rpds-py==0.26.0
 
rsa==4.9.1
 
shapely==2.1.1
 
six==1.17.0
 
sniffio==1.3.1
 
SQLAlchemy==2.0.41
 
sse-starlette==2.4.1
 
starlette==0.47.2
 
tenacity==8.5.0
 
typing-inspection==0.4.1
 
typing_extensions==4.14.1
 
tzdata==2025.2
 
tzlocal==5.3.1
 
uritemplate==4.2.0
 
urllib3==2.5.0
 
uvicorn==0.35.0
 
watchdog==6.0.0
 
websockets==15.0.1

Werkzeug==3.1.3
 
zipp==3.23.0
 
firebase-admin
 
groq
 
structlog
 
supabase
 
openai
 
langfuse==3.3.5
 
openinference-instrumentation==0.1.38
 
openinference-instrumentation-google-adk==0.1.4
 
opentelemetry-instrumentation
 
opentelemetry-exporter-otlp-proto-http==1.35.0
 
posthog==5.3.0
 
mem0ai
 
aiohttp
 
aiofiles
 
pydub
 
gtts
 
pyttsx3>=2.95
 
openai
 
ffmpeg
 
imageio_ffmpeg
 
google-cloud-texttospeech
 
pydub

aiohttp

aiofiles

pydub

gtts

pyttsx3>=2.95 

openai

ffmpeg

imageio_ffmpeg


google-cloud-texttospeech

pydub 
aiohttp 
imageio-ffmpeg
 
boto3
 
httpx
